Update recorder.cpp
Expand getNextAction method to accept an additional, alternative iValue to match called iValue2. This is necessary because midi events generated from keyboards generally send the note off event with a velocity of 0x40, regardless of the note-on velocity. This change allows getNextAction to find these note-off events correctly.

The change in line 557 also eliminates a second and pointless check for iValue == 0 in the original code.