Update gd_mainWindow.cpp
implement song meter and musical breaks; These additions place the songmeter on the main window and make some adjustments in the window layout. The adjustments are making room for the songmeter and an enlarged timing group. I have also move the menu down to begin at y=8 instead of -1. I don't know what it was doing at -1, it may have made it look more like a menu ( possibly ) but it also meant the half the button areas are where outside of the Fl_group they were a part of. This mean that half the button space was dead as far as clicks were concerned, it drove me nuts!