Update kernelMidi.cpp
implement midi keyboard echo; when a channel has both midiTunnel and midiOut true, but the sequencer is not in record action mode, then echo the midi keyboard events directly to the midi output so you can hear the notes before recording. At the same time swap out the midi channel in the incomming message for the true midi destination channel for that MidiChannel.