Update midiChannel.cpp
implement songmeter and musical breaks; implement midi stop() functionality which wasn't necessary before. We use the start() and stop() methods to implement musical breaks by turning this channel off when it's not being played ( turned off by actualBreak advancing), and back on when it's time. As a side effect, when using musical breaks in the song meter this means the midi channels are automatically started, they don't need a manual "start"

Also, change the operation of midichannel start() to send the program and bank change messages anytime we hit the play button on a midi channel and the midi output is enabled on that channel. This is better when using the midi keyboard echoing, now if we change the program select and hit the play button we'll get the new voice immediately.