Major MidiMsgFilter redesign
Lambda filters have been removed. It is impossible to serialize them, which is a good enough reason to give up on them.

Instead, a few operators have been overloaded, namely &|! for and / or / not manipulation of filters. Note that these actually create new, fully compatible MidiMsgFilters that can be used later. For example:
const MidiMsgFilter MMF_NOTEONOFFCC = MMF_NOTEONOFF | MMF_CC;
Believe it or not, it actually works :D
And all that is json serializable!
Unfortunately, this also means that manipulating filters will be seriously limited.

As a bonus, a "check message" operator have been introduced, like so:
(MMF_CC << mm)
Returns boolean result of checking whether MidiMsg mm is a CC message. How cool is that!